15983 FIX time_period: verify time_period usage before deletion and fix the reference for time_periods being saved to the configuration file in exclude field
Prior to this werk, the REST API did not correctly check if a time period to be deleted was being referenced by another object. This werk fixes this and prevents the REST API from deleting a time period that is currently being referenced.

Also, prior to this werk, the REST API was incorrectly storing the reference to other time periods in the exlusion field. This werk fixes this and now the REST API stores the correct values in the configuration files.

Ticket: SUP-15073

Change-Id: I2c90141a0421d5a6ac66e7193be080fc771cf76e