15194 SEC Fix command injection via RestAPI / Password Store
NOTE: The Werk in this commit is backdated to 2.2.0p4 even though this
commit is published with version 2.2.0p8. The reason for this is that
the vulnerability described in this Werk was already fixed with commit
e3396685b29 / Werk #15889.
This way, this Werk should show up for the version that first fixes the
vulnerability.

In addition to adding the Werk, the commit adds the sanitization
improvements consistent with the other branches. See Werk text for
further info.

We have assigned CVE-2023-31209 to this vulnerability.

CMK-14149

Change-Id: Id7087d6d57e6fc62f01dff8543737f880740e676