15194 SEC Fix command injection via RestAPI / Password Store
Prior to this Werk, users with the permissions to (a) use the RestAPI, (b) create passwords in the password store, and (c) create active checks were able to run arbitrary commands on the site.

This issue was found during internal code review.

<b>Affected Versions</b>:
LI: 2.0.0
LI: 2.1.0
LI: 2.2.0 prior to version 2.2.0p4

Note that at the point of publishing this Werk and fix, the current version 2.2.0 was already not affected by this issue anymore, as the issue was already mitigated by Werk #15889.

<b>Indicators of Compromise</b>:
Check the password store for passwords with unusual identifiers, review add-password events in the audit log.

<b>Vulnerability Management</b>:
We have rated the issue with a CVSS Score of 8.8 (High) with the following CVSS vector:
<tt>CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:H/A:H</tt>.
We have assigned CVE <tt>CVE-2023-31209</tt>.

<b>Changes</b>:
This Werk adds proper sanitization of the affected parameter on core commands.

CMK-14149

Change-Id: Id7087d6d57e6fc62f01dff8543737f880740e676