15971 FIX Licensing: Missing license transfer in distributed setup
Prior to this werk, the license of a central site in a distributed setup would not be transferred to remote sites if the remote site was in the "Free" license state, as this state is not allowed in distributed setups.
This meant that remote sites could effectively not be licensed once they entered the "Free" state.

Now the license information is distributed to all remote sites when the licensed verification is performed.

CMK-14163

Change-Id: I51da2694acf57d4e7240373eb59b87ab08e446d4