15908 FIX df: Showing wrong WARN/CRIT levels when magic factor is 1.0
When levels for free/used space were configured and the magic factor was configured to 1.0, the df check-plugin was using the default parameters as WARN/CRIT levels.
This behaviour differed from 2.1 and causes the check-plugin to sometimes report an incorrect state.
The problem has been fixed and the behaviour is the same as in 2.1.

SUP-15120

Change-Id: Id1c2e2251bf82a00afd933fc36fba0180a947ffd