Move error handling out of check engine
Error handling is not part of the check engine but of the caller.  For
example, the check engine shouldn't know anything about our crash
reports.  It should crash and the caller create a crash report when
appropriate.  Generally, error handling is external to the component
that produces the error.

CMK-12298

Change-Id: I2be2eaedf3ae6b946d5d0db393b7b2cb637e156f