aws: rename assume_role to access
One of the following commits will add another element to the
"access" Dictionary, which does not fit in the "assume_role" identifier.

The old migrate code ("piggyback_naming_convention":
"ip_region_instance") was removed deliberately, because it's already
present in 2.2 and thus was already executed during the update process.

SUP-14992

Change-Id: I2b323ba89d8fc19ee7d1f78d8b836446e81cd461