15985 FIX AWS agent handles wafv2 rules with ByteMatchString correctly
When a wafv2 rule with the ByteMatchString was configured the agent would crash. In the debug output a message about an unserializable entry would appear.
In the AWS UI you can configure a rule with ByteMatchString in any field using "Search string".

This has been fixed now. No action required.

JIRA: CMK-14298
Change-Id: I2fb977c87bcc4e92dfb1e65dc65cc9f11f4f3363