16030 Alphabetic sorting for Kubernetes dashboard elements
The elements of the cluster overview dashboard were not sorted at all previously.
They are now sorted alphabetically.
The sorting of the cluster and namespace dashboards was partially done on host name, which leads to a sorting by namespace. To unify this behavior, all list elements are now sorted by name alphabetically

Change-Id: Ice4efc7ca5551967ed6cd7b656bb6d0a50cae2fb