15260 FIX Fixed rescheduling of non-existent service
When a Livestatus query with a "WaitObject:" line for a non-existent service
was sent to the monitoring core, a crash happened.  This could e.g. be
triggered by direct Livestatus queries or via a "Reschedule" in the GUI
under certain circumstances.  This was a regression in 2.2.0 and has been
fixed.

Change-Id: Ia43f4b40b0247d1670cc7f0a58f18a83e89ff1f3