15292 FIX Changelog will only contain Werks from current major version
Checkmk has a built in changelog viewer that shows Changelog items
called "Werks". For previous versions of checkmk it showed "Werks" for
all versions of Checkmk.

It turned out that the changelog was not complete: For previous major
versions of Checkmk not all Werks were displayed. As there is little
benefit to display Werks of previous major versions of checkmk we
decided to remove those.

The whole list of all Werks is available on our homepage:
https://checkmk.com/werks

When updating to a new major version, the update process will check if
there are unacknowledged werks and tell the user about this, as this
list will no longer be available after the update.

CMK-13658

Change-Id: Id5ded1a27fa536c5801dcee40c1b2dd9fb588d8a