Remove integration test for number of werks shipped
The tests made sens as previously the number of werks were
monotonically increasing.

Since Change-Id=Id5ded1a27fa536c5801dcee40c1b2dd9fb588d8a we expect the
number of werks to be zero when we release a new version of Checkmk and
increase the major version of Checkmk.

Change-Id: I3edeb31007c6ba2821594ad446c8a6ebeebf30b4