Fix docker omd update in docker
In

Werk 15292
Change-Id: Id5ded1a27fa536c5801dcee40c1b2dd9fb588d8a
commit 5eed710c20a8efe7073b4f1678c0429a256387c4

we assumed that we have always access to the omd installation of the
previous version. This is not always the case: When `omd update` is
executed in docker, only the current version is available. In this case
we skip warning about unacknowledged werks as we have no way to
determine them.

Change-Id: Ic0aafe3ff69dad9899cede5a8abf756d1bb4109f