16096 FIX authentication: allow rest-api access when gui access is restricted
This werk addresses an issue with regards to rest-api access. When the
distributed monitoring config option

"Users are allowed to directly login into the Web GUI"

was disabled, we inadvertently blocked access to the rest-api as well.
We now allow rest-api access while still blocking access to the Web
GUI in this scenario.

Not pretty but works. check_auth() is called before we do any url matching,
hence we don't know at this point if the request is destined for the
rest-api or not.

SUP-15133

Change-Id: I7ca0640131908f2962a0e2802d1b9c3b60e1310e