Let mypy check all of _validate_parse_function(). Comments.
Mypy thinks that a parse_function can *never* be a types.FunctionType,
so it considers most of _validate_parse_function() unreachable code,
effectively disabling type checking it. This is due to the type of the
parse_function argument (effectively a union of Callables) and the
implementation of types.FunctionType plus a few ugly technical
details. We circumvent this by calling inspect.isfunction(), which
does the same, but this is hidden from mypy. :-}

In any case, we probably just want to check for callable(), but this
would be a separate change.

CMK-14437

Change-Id: Ia403b3f062459c60a8c9c6841533800fbe69f79c