docker tests: Fixed local execution
So far, "test_build_using_local_deb" removed the DEB package and
only recreated it in a CI environment, which prevented local
execution of the whole suite. This change fixes this by creating
a backup file and restoring it at the end.

I also refactored "test_build_using_local_gpg_pubkey" slightly so
it is more similar to "test_build_using_local_deb".

Change-Id: Ia4c6a0abc65af4ace670f98b96bc2a6b495cdfb6