16098 FIX notification rules: match_folder value now matches correctly to available folders
This werk addresses an issue when creating or updating a notification rule
via the REST-API.  Previously, it would not correctly match the desired folder
and as a result, you had to set this option to disabled which meant all rules
were created in the main folder only.

Now the match_folder value field should match a folder id, in the form of the
folder path, replacing / for ~

E.g.

/folder1 ->  ~folder1
/folder1/folder2  ->  ~folder1~folder2

Change-Id: If58bcc32bc74bcb485bc66117100dd32fecc348d