16125 Remove Python 3.4 as an option for Windows Agent
Older windows versions (e.g. Windows Server 2008 RC2) do not support Python versions newer than 3.4.
For these systems we offered the option to install Python 3.4.4 as a Python environment.
This environment was then used for various agent plugins.

Since Python 3.4 reached its end-of-life in 2019 and Windows Server 2008 RC2 its end-of-life in 2020 this support is now removed.
During the update you will receive warnings if you have rules to deploy Python 3.4 and these rules are change to not install Python anymore and to use the system Python instead.

If you want to continue to run and monitor these systems you can still do that but if you use agent plugins that require Python you are now responsible for providing that environment.

Change-Id: I0f2d1a12b30b9fc6e2411840f30eeadf90f5c20a