Cleanup one GUI import from base unit tests
Moved the UI test to the cmk.gui test hierarchy and separated the shared
data structure into a tiny helper module. The location of the helper
module is still questionable.

Nevertheless, this is already a reasonable improvement, since it serves
the purpose of removing the cmk.gui imports from the cmk.base unit
tests.

Change-Id: I77ce576d438c84d9236a8ebec4218cd541a858be