Prevent cross-component imports in unit tests
The test code is not special compared to the regular program code in
regards to layering. It also need to respect the rules we have in the
other program code.

So far all tests were completely excluded from the layer checker, which
was only meant as a temporary exclusion to get started with the checker.
Turns out it was not so temporary.

This change now at enables the layer checker for `tests/unit`.

A few local suppressions are still in place, at least this is all local
now.

Change-Id: I0c1e5e2045ad37af700283be5fa61c9872840d69