16126 FIX Ignore invalid session cookie for header-auth
When an user logged in via HTTP header authentication (e.g. the legacy SAML method with mod_auth_mellon) and that user had a session cookie from a previous session which is now invalid the creation of a new session was aborted.
This Werk fixes this so previous session cookies are ignored if they are invalid and a new session will be created.

Change-Id: Ie5aea39d9af524429b2243a4a3d531420e138d98