15774 FIX checking: Prevent data loss on hosts with longer check intervals
Previously, the validity of agent data for discovery and HW/SW inventory checks was 2 minutes.
When checking is performed with a longer interval, the discovery would retrigger fetching of the agent data.

In some checks this leads to data loss, because the checking fetches the data again and the data
fetched by the discovery is lost.

The cache interval for discovery and HW/SW inventory checks is now 1.5 * checking interval.

SUP-15688

Change-Id: I33a1d41a5e425124c26e7355a56e2d1f3dc9a79b