15643 FIX nvidia_smi: Show Use full PCI bus ID
This change affects users of the following checks:
* nvidia_smi_memory_util
* nvidia_smi_gpu_util
* nvidia_smi_en_de_coder_util
* nvidia_smi_power
* nvidia_smi_temperature

Previously, the PCI bus ID of NVIDIA GPUs would be incorrectly removed. This caused GPUs to be
omitted, which should have been discovered. With this Werk, the full PCI bus ID is used as a service
item. Users must rediscover the services.

SUP-15712

Change-Id: Ib4e13eff8d395c4b4b1da9c3990cbf1159bf2900