16135 FIX apache_processes: we now replicate the configuration on remote sites
This werk addresses an issue where we weren't replicating the configuration
for the number of apache processes.  Now we do.  On activate changes the
number of apache processes configured on the main site will now be replicated
on the remote site.  An apache restart is required for these changes to
take effect.

SUP-14333

Change-Id: I55197f0b821c54a74405c3efc39ffe8f09c4ae22