15293 FIX mk_oracle: fix custom sql sid matching
Configuration variable `$SQLS_SIDS` can be used to define a comma seperated
list of sids for which the custom sql is executed.

Before this werk the custom sql was executed if the sid was a substring of one
of the elements of the list in `$SQLS_SIDS`.

Now custom sql is only executed if the sid is equal to the whole element
provided in `$SQLS_SIDS`.

Change-Id: I758f396dfccc7441bb7fbefb1d843d1d4ba35057