Fix get_bloat for Windows and use PGPASSFILE
This commit fixes two issue.
1. The get_bloat executes psql for each database.
This is necessary because otherwise it exceeds the maximum 8191 characters limit of windows https://learn.microsoft.com/en-us/troubleshoot/windows-client/shell-experience/command-line-string-limitation
Which triggers the error "The command line is too long."
I observed it with 3 databases

2. The PGPASSFILE Env varaible is only overwritten by the instance pg_passfile when it is not empty.
In this way it is possible to use a system wide PGPASSFILE evnironment variable.