16161 FIX discovery: Fix missing datasource error for cluster hosts
Previously, a missing datasouce error would be shown in a discovery
page of a cluster host if the cluster host had 'No IP' configured
but had nodes with an agent datasource.

Clustered services of a cluster host would also become vanished.

Checkmk now supports the described setup, there is no error in the
discovery page and clustered services are discovered.

SUP-15433

Change-Id: I0d1a949603467b9bb72e560e652e7e6777369c31