16161 FIX discovery: Fix missing datasource error for cluster hosts
Previously, a missing datasouce error would be shown in a discovery
page of a cluster host if the cluster host had 'No IP' configured
but had nodes with an agent datasource.

Clustered services of a cluster host would also become vanished.

Checkmk now supports the described setup, there is no error in the
discovery page and clustered services are discovered.

SUP-15433

Change-Id: Ia0ce5b354c2a1bcd8f33bd660b7c7a89db9d28c6