15570 FIX Handle redirect RoutingRules of azure's application gateways
If you're having a http redirect routing rule configured in your azure application gateway, the parsing of the section may have failed with:
C+:
pydantic.error_wrappers.ValidationError: 10 validation errors for AppGateway
C-:

This is due to the fact, that a redirect routing rule may not have the previously required information (backendAddressPool and backendHttpSettings).
Those fields are now not required anymore.

Change-Id: I6fab9065100c54d06488091c533ddafaafa911da
JIRA-Ref: SUP-15950