15098 FIX setup form page rendering stuck when page has lots of select boxes
Previously, when a HTML form page was rendered with a lot (>hundreds) of
select boxes, the augmentation process to make them searchable was taking
a very long time.

Until a proper fix for this performance problem can be found, we now prevent
the augmentation process whenever it would take more than 3 seconds.

The user-interface is almost not affected, only the select boxes are no longer
searchable. No manual action of the user is necessary to make this work.

Change-Id: Ic0d81df11be84289476876801d0c1ba00f331b1b