Unpin astroid.
As a consequence of this unpinning, we get the new pylint 3.0.1.
This in turn requires a few tweak caused by API changes for checkers.

In addition, the cyclic-import checker works again, even with --jobs.
The really scary part is that we get quite a few cycles reported.
We'll need to investigate this further, for now we just disable the check.

Change-Id: I892657f27f6db1c74072ee4199f4ed14efe3df1c