15977 FIX db2_backup: Correct timezone difference for last backup date
This werk is relevant for users monitoring the age of the last backup time of db2 databases in different timezones.

The date/time of the last backup of a db2 database is currently stored in local host time without the information about the host timezone. When this time is used to check the age of the last backup, it is interpreted in the Checkmk server timezone.
When using different timezones, this leads to incorrect values for "Time since last backup".

The db2 agent plugin will now store the time in UTC and the db2_backup check will interpret the time accordingly.

You will need to update the agent mk_db2.linux to receive the corrected times.

Change-Id: I07db12befcc96d5c35ffd078dbd3cb184a67dce5