Skip syncing not existing async-timeout directory
Since Python 3.11.5 async-timeout is not built as a dependency of
redis-py, because it declares the "python_full_version <= '3.11.2'"
marker. Seems our build mechanism is not aware that it is being
skipped for this reason.

Try to skip the failing rsync for now. We'll need to find a better
solution for this next week.

Change-Id: I1af3ac9784454cd88d769750fffd917f58542b17