Fix centos-8 build: wrong 'as' command used
install-gnu-toolchain.sh creates the symlink /usr/bin/as pointing to
/opt/gcc-13.2.0/bin/as. Later on install-cmk-dependencies.sh installs
the binutils package, which replaces /usr/bin/as with the packaged
executable.

With this change we now create symlinks pointing our GNU toolchain to
/opt/bin in addition. This should have priority over the files in
/usr/bin, if the PATH is set correctly in all necessary places (let's
hope).

This affects all distros, which might also have used the OS tools. But
for them it might have been no issue, in case the tools are recent
enough.

As a follow-up, we need to move everything to /opt/bin - but that needs
separate testing.

Change-Id: Ifba67ebc261ffa411ca2adcfabf0677cd9bf17d2