remove pytest fixture, use function instead
as soon as you set the environment variable `TZ` to something, the
skipped test started to fail. the problem was, that the fixture (parsing
the section) was outside the `with on_time` block setting the timezone.
So the section was parsed with the timezone of the system which is
randomized when executed via `make test-unit`.

Change-Id: I740a7ea8cd83c9680fb355a3e5eb8ea71bfc17f1