16106 KUBE: Addition of support for Kubernetes version 1.27
With this release of Checkmk, we introduce support for version 1.27 of Kubernetes.
In Checkmk 2.3, support for Kubernetes version 1.22 is removed. The supported versions
are listed below:

LI: Checkmk 2.2: 1.22, 1.23, 1.24, 1.25, 1.26, 1.27
LI: Checkmk 2.3: 1.23, 1.24, 1.25, 1.26, 1.27

The list of supported versions may not apply to future patch versions. For such cases, a
new werk will be released.

CMK-13673

Change-Id: I030864767b9dd377543b840e8653d194cda0e97e