15100 FIX re-use of context in multiple threads in activate changes
When running "activate changes", a "context" object is created and
passed to multiple threads. The concurrent use of this context is
not supported and can lead to crashes.

As a fix, now every thread receives their own copy of the context
leading to complete isolation of the threads.

No further action is required to activate the fix.

Change-Id: I078040d6a00faae61c7d72d55eb0863597812395