prediction/check api: pass prediction updater as callback
This patch changes the way the predictive levels are
computed during checking.
Instead of relying on two global variables we pass
a callback. This makes the de-facto dependencies of
the backend explicit: livestatus and cmk.utils.prediction.

I don't particularly like the magic key in the parameters,
but I think this is better than the alternative.

This solution relies on the fact that the parameters are
typed Mapping[str, Any].
In some future API we could either find a better solution
altogether, or at least expose some Protolol that describes
what we expect from a "PredictionUpdater".

Change-Id: I47749ec781e7563ba05ddfedcd9fde907fee9842