15299 FIX nfsiostat: fix units and parsing
Units falsely displayed seconds, although milliseconds are corret.

Newer versions of nfsiostat output more data, this confused our parser.

Before we parsed just the first 16 numbers: 2 global, then 7 for read
and 7 for write. But nfsiostat will now report 22 numbers: 2 global, 10
for read, and 10 for write, so the first 16 numbers have a total
different meaning now. We now parse the first 2 numbers, and then 7
numbers after "read:" and 7 numbers after "write:", so we are compatible
with the old and new format with just one parse function.

Change-Id: I6a3b467f1f5d0ac7a2e343ebd55e3af1b5e67465