don't discard all mails if mail_id's are bytes instead of str
Depending on the mailserver + python version, the mail_id can be bytes instead of str. i.e. in 2.1 with the older Python 3, out onprem exchange response was parsed to str, in 2.2 it is parsed to bytes and hence all mail_ids were discarded here and check_mailbox always pretended no old mails were found and the check was essentially functionless. However, fetch_timestamp can totally handle mail_id's in bytes format, so we can just pass those as well in the next line :)