16159 linux-bonding: introduce check for expected bonding mode
The service monitoring rule "Linux monitoring service status"
was previously being used for both Linux and OVS.  This werk
now splits this rule in two.  We now have the following rules.

Linux bonding interface status -> Linux
Bonding interface status -> OVS

The reason behind this, is that we have now introduced an
expected mode configuration parameter only for the Linux
rule. So you are now able to select the expected bonding
mode and the state you want to display when this condition
is not met.

SUP-14636

Change-Id: Iadb06a9fc1ec1dae7cb4c87268ca00abd1e72b23