15836 FIX relax memory restrictions for fetcher process
In previous releases, an overly stringent memory limit for the fetcher
process could result in frequent restarts and an excessive number of
entries in the cmc.log file.

With this release, we have addressed the issue by adjusting the memory
constraints. The fetcher process now has an increased memory limit,
which should significantly reduce restarts and log verbosity, leading
to improved stability and performance.

Change-Id: I0ad5223902c668de3e49b7eec78bb4621e9e4d05