16216 FIX Resolve Runas section in Checkmk Linux agent
Prior to this werk, the Runas section of the Checkmk agent
contained a wrong trailing slash which resulted in an error
with the find command. As a direct consequence, some or
potentially all Runas rules were inadvertently ignored, leading
to the plugin, not intended for root, being executed with root
privileges. This werk fixes this behaviour.

Change-Id: Iaa3ace9b4d1eb1c3e554826dc537d53be98ebe84