16215 FIX agent_elasticsearch: resolve error when requesting indices stats
There is a special error case which arises when trying to request the
indices stats from Elasticsearch which has at least one alias
pointing to a closed index. In the previous version, the special agent
completely failed in such scenarios. This werk resolves this by
requesting only the accessible indices.

SUP-16295

Change-Id: I679a2f777f2c4584019b9f84f5d800478f68ec47