Always unset variables when doing exec to other OMD version
The exec to another OMD version can always also mean a change change of
the Python version. So it is never correct to hand over the PYTHONPATH
or LD_LIBRARY_PATH from the origin version.

This is independent from the OMD command being executed and not specific
to the "omd update" command.

Change-Id: I8ec91ef220f821e13ad5428afc50f2b69f5d112e