16219 SEC Limit length of Hostname
Prior to this Werk it was possible to create Hosts with arbitrary length.
Since Checkmk stores information in files which paths contain the hostname these path could exceed the allowed length leading to various errors to an extend that rendered the usage of parts of the GUI useless.

We found this vulnerability internally.

<b>Affected Versions</b>:
* 2.2.0
* 2.1.0
* 2.0.0

<b>Vulnerability Management</b>:
We have rated the issue with a CVSS Score of 2.7 (Low) with the following CVSS vector:
<tt>CVSS:3.1/AV:N/AC:L/PR:H/UI:N/S:U/C:N/I:N/A:L</tt>.
We assigned CVE-2023-23549 to this vulnerability.

<b>Changes</b>:
This Werk adds a maximum length of 253 characters for the hostname.

CMK-15105

Change-Id: I4be4745ede4220d05f4ff01d51aa6252dc9a1b57