Profiling middleware: allow for profiling the application startup phase
 * Introduce ProfileSwitcher middleware
 * Adapt profiling tests to reflect new calling convention
 * Adapt index.wsgi to use the new structure
 * Remove old middleware from Flask app
 * Update Pipfile: remove no longer necessary repoze.profile
   middleware, make pyprof2calltree am explicit dependency.

The ProfileSwitcher middleware can now import the wrapped app
at runtime, so that importing and initialization can be profiled
as well. To achieve this, a default "config" is used for the
very first request, when no config could have been loaded yet.

For the first request, it is intentional that triggering the
profiling middleware via a "_profile" GET query-string parameter
is possible.

The subsequent requests will honor the configured profiling behavior.

CMK-14511

Change-Id: I33514e6395bcfb4849c421daa9647c944fffe319